Better cleanup of failed node up handling
When a node comes back up, we try to create a new connection
pool for it.  If that process fails, listeners and the
control connection need to be notified.  Prior to this change,
they would not, so the pool could remain in a shutdown state
and the control connection could be unaware that its connection
had failed.