Use WeakValueDictionary for prepared statements
When nodes are added or come back up after being down,
all known prepared statements are re-prepared on them.
In order to avoid preparing statements that will never
be used again, we'll use a dictionary of weak references
to the statements so that once they are GC'ed, they are
removed from our cache.