Use same conn to check schema agreement
Before doing this, the control connection would be used to check
for schema agreement.  The schema check could occasionally complete
before the node that the control connection was open to had even
registered the schema change.  By re-using the connection that the
schema change was performed on, we avoid this race condition.

Fixes PYTHON-33