Use select backend for libev
It appears that the epoll backend may have been causing the hanging
write requests at the connection level, which appears to be the last
major cause of PYTHON-27, PYTHON-29, and PYTHON-32.  For now, we will
hardcode select as the libev backend; this would be the best choice in
almost all circumstances, in any case.