Undo param passing for non-prepared statements
There are a couple of reasons this can't be supported at the moment.

The first, and most serious, is that the passed parameters must be in
their binary form, not a string representation.  For non-prepared
statements, the driver doesn't know what Cassandra types it's dealing
with, so the best it can do is map python types to string formatting
functions (see cql_encoders).  It can't tell if it should convert 123 to
the Int32Type binary representation or the IntegerType binary
representation.  (Likewise for FloatType and DoubleType.)

Second, Cassandra doesn't accept a map of parameter names to parameter
values, so even though there are named placeholders, the driver would
have to parse the query string to handle a dict of parameters.