Fix duplicate node-up handling
a22f80b8 introduced a bug where multiple UP notifications
could result in multiple reconnectors being started. It was also
possible for the executor threads to become deadlocked waiting on a
condition that would never be set, which meant that future down and up
node handling would never be executed.