Trivial: fix typos; improve docstring example
It's not explicit that the alternate function accepts all the arguments
of execute_concurrent until looking at the source code. I think that
adding the concurrency=50 argument to the example in the docstring
satisfies this.

My first idea was to explicitly name the concurrency and
exit_on_first_error args, but then we'd have to maintain two functions
in the future instead of of one.