cqle: Update invalid inet test with a string that won't parse
Cassandra seems to be accepting various bad strings and inserting
'92.242.140.2' (for many different strings). Need to follow up on that.
Meanwhile, this string does induce an Invalid request.

Side note: I'm not sure this test is worth having, unless there is
supposed to be some kind of local validation. We're not testing the
server here.