First attempt at fixing previous value tracking.
I reused the global '_is_persisted' flag on models to tell the value
manager about the origin of the previous value. Previous value is kept
aligned to the provided value only if the model instance is persisted,
else it is reset to None.

If I succeed, I propose to rename 'previous_value' to 'persisted_value'
then.