Fix hang when using eventlet and ssl
The eventlet read loop runs select before each recv call. This will
fail because recv() on an ssl socket might read more data than
requested from the underlying socket in order to have enough context
for decryption. These extra bytes would be returned by the next recv
on the ssl socket, but are not known to underlying socket so select
will hang.

It is my understanding that eventlet implements blocking IO by using
poll under the hood and yielding, so the select should not be
necessary at all