Don't use a kwarg to pass a callback to weakref.ref
Turns out that until recently, weakref.ref() silently ignored its kwargs, but
a recent Python patch causes it to throw an exception instead. This makes cqlsh
error out immediately when using it with the latest Python version.

In order to correctly register the weakref callback, pass it as a regular
argument instead of using a kwarg.

For more information, see this Python issue and related Python 2.7 commit:

    https://bugs.python.org/issue17765
    https://hg.python.org/cpython/rev/60de9c6188cc