add _CANCELLED sentinel for perma-cancelling timer
This does not eliminate, but it shrinks, the window for the race
condition that caused PYTHON-644. In this race, _set_result could be
called from the event loop before send_request calls _start_timer.

The goal of the _CANCELLED sentinel is to allow permanent cancelling of
all future timers for a given ResponseFuture. Unfortunately, there's
still a race -- if _timer is cancelled during _start_timer.