Fire callbacks on ResponseFutures before unleashing result()
When callbacks and blocking on .result() are mixed, it's currently a
race for which happens first. By flipping the order of operations here,
we guarantee that the callbacks are all complete before unleashing the
main thread. This makes it possible to reliably do stuff with callbacks
like collect metrics while the main application uses .result().