Do not error on null host_id entries
SNI compatible servers may be running in very dynamic cloud
environments. We are running our DSE servers on a dynamic kubernetes
environment, where operational maintenances occur very regularly. This
is leading to "null" entries in the host_id tables, i.e.:

```
ibm@cqlsh> select host_id from system.peers;
 host_id
--------------------------------------
 ed7b1b26-7cad-47a2-9ef4-638360127526
 415e33c0-a9cd-4989-a727-40adef5d38d9
                                 null
                                 null
```

The DSE servers usually clean these null values up themselves after some
time, but while they still exist, the python-driver is throwing throw an
error and failing to connect.

This PR changes that by simply logging the existence of the null value
in the system.peers table. We argue that there is no need to throw an
error and fail to connect to the service, since the caller of the
changed method loops over all the values and handles the valid values
later.