Don't create Cluster with hosts if having cloud config
We set out to build a Flask app that talks to Astra.
The CQL plugin we used turned out to be crashing if we didn't give it hosts to connect to (https://github.com/thegeorgeous/flask-cqlalchemy/blob/master/flask_cqlalchemy/__init__.py#L48).
But we also had to use the `cloud` key to pass in Astra credentials (via `setup_kwargs` nearby the link above).
This exploded on us in cqlengine.connection#setup() where the hosts passed into Cluster() made that constructor blow up because:
`ValueError: contact_points, endpoint_factory, ssl_context, and ssl_options cannot be specified with a cloud configuration`

We're not sure this is the right place to patch, but it helped us at least. We figured coming out with a patch directly, instead of an issue, might save some time.