HostConnection: implement excess connection keeping
When assigning shards to connections, Scylla chooses the shard with the
least number of connections. If the pool misses a single connection to
some shard which is not the shard with the least number of connections,
the pool will not be able to connect to it because the attempted
connection is immediately closed in case it gets the wrong shard.

In order to make this process more robust, now the pool can keep
"missed" connections in the excess connection pool. The excess
connections are not used to serve requests, they are kept only in order
to affect the algorithm on the Scylla side which chooses shards to
connections - and make it possible to connect to shards which are
currently not the least loaded with connections. When the pool becomes
full, excess connections are closed. Furthermore, the number of excess
connections is bounded by 3 * shard count; if it goes above this number,
all excess connections are closed.

This change is also necessary for the logic of replacing of connections
with too many orphaned stream IDs - which is coming in later commits of
this PR.