tests: fix test_bad_contact_point
since recent changes to metadata, on first round we remove all
host that we creating with unkown host_ids, so this test is failing
cause it's expects all the corrent  hosts to be available in the,
metadata, they would be available, but they were missing.

seems like we didn't update the peers with thier newly found host_ids
and hence they were remove (only a new phase of `refresh_node_list_and_token_map()`
would add them back to the hosts list), now that we update them
same as we update the local (control_connection host), test is working.