Fix code incompatibilities with Python 2
There are some functionalities used that don't work with Python 2.
- Format strings
- list.clear()
- Type hints
- super() without arguments
- Packages without __init__.py file
- Some import names (futures.thread -> concurrent.futures)

Import behaviour changed between Py2 and 3 and one file was missing
`from __future__ import absolute_import` line (that makes
the behaviour consistent between 2 and 3) which caused import error.

Some members in "c_sharding_info.pyx" had "str" type. This type maps to
"str" both in Py2 and 3 - and those are different types - raw bytes in Py2,
unicode string in py3. This caused errors in Py2, because code was trying
to assign unicode strings to those members.
The fix is to use Cython's / Pyrex's "unicode" type - it maps to
"unicode" in Py2 and to "str" in Py3.

This commit fixes all of those problems.