Update entrypoint and host_id only when `SniEndPointFactory` used
Since there are use cases when the user uses external address (i.e.
not the node broadcast address), and not using `AddressTranslator`)

the assumption of the user is that it would be able to connect
and use the control connection without creating any more connection
to other nodes.

when we chnage the logic to work based on host_id, we decided to
remove the initial control connection host, since it didn't had a correct
host_name, which break this use case.

so for now we'll leave the removal only for case we are sure it's
the expect thing to happen, i.e. when `SniEndPointFactory` is used.

Fix: #184