PYTHON-1267 Add a more useful error message for NaN support.
Ideally we would be able to have full NaN support in the cassandra driver for python, as python Decimal type supports NaN.
This isn't really doable though as Cassandra line protocol currently (v5) doesn't support it.  This gives us a better error message
so at least it will be easy for users to understand what is going on when they try to serialize a NaN-valued Decimal.