Fix wait_for_schema_agreement deadlock
Fixes https://github.com/scylladb/python-driver/issues/168

Fix works by extracting part of on_down that marks host as down
out of the executor - so it does not need to wait for free thread.
When host is marked as down, wait_for_schema_agreement can finish,
which in turn enables rest of on_down (the part that still runs on
executor) to be executed.