pool: log error when failed to connect to shard
The exception from `HostConnection::_open_connection_to_missing_shard`
during connection failure is silently dropped by the callers.

This function is submitted to the `ThreadPoolExecutor` which
assigns the result of this function to the future
(either success or exception).

The callers throughout the code ignore the future's result
and that is why this exception is ignored.

In this commit we add an error log when opening a connection to the
specific shard fails.