Add setting connected_event flag in libevreactor
Before, the connected_event flag was set in every implementation
of Connection but this utilizing libev. Other reactors have the
same `self.error_all_requests(ConnectionShutdown(...))` logic, but
they have `self.connected_event.set()` after that, so it was probably
an oversight (copy-paste mistake?) that it was missing from this
reactor. That was causing the driver to sometimes hang for >3 minutes
when shutting down.

This commit adds setting the `connected_event` flag in `close()` in
`LibevConnection`.