Only add host if endpoint is not already present
In d735957 there was a functionality added to reresolve
hostnames when all hosts are unreachable. In such a scenario,
the driver will try to save the situation by reresolving the
contact points in case it helps.

However, if there was no ip address change, this results in
creation of new (duplicate) Hosts (same endpoint different host_id)
which in turn starts new reconnection processes. Those duplicate
reconnection processes can make the situation worse when the driver
regains connectivity with the cluster. Different Hosts with the same
endpoint are reconnecting in different moments and this cause host
to be unreachable in unpredictable moments.

This commit introduces checking if the resolved endpoint is already
present in Cluster Metadata information. The new host is added only
if this condition is not true.

Fixes: #295

Refs: scylladb/scylladb#16709, scylladb/scylladb#17353