Remove endpoint to host_id mapping when removing host by host_id
To remove host not found in peers metadata remove_host_by_host_id is used.
In most cases we want to remove host that is a duplicate of host found in
peers metadata with the same endpoint but different host_id. Because of that
mapping in _host_id_by_endpoint is already overwritten with new host found
in peers metadata so we don't want to remove it.

In case that we want to remove host that do not have its duplicate with
different host_id in peers metadata we do need to remove mapping from
_host_id_by_endpoint. This commit intruduces handling this case.

Refs: https://github.com/scylladb/scylladb/issues/17662