AsyncioConnection: fix initialize_reactor when called in event loop
Previously, if executed within existing asyncio loop, driver would take
the loop, assume it's not used and start it in a separate thread.

Additionally, if executed outside of loop, driver would create a new one
and make it default for calling thread.

Those behaviors are wrong so they are changed. Now driver creates its
own loop and executes it in a thread. Code that handled pid changes,
which can happen when class is transferred using e.g. multiprocessing,
is fixed too - previously it didn't create new thread after such transition.