CI: download libev via conan, for windows builds to have it
windows builds so far was running with having libev available
and until this sync the fallback for python 3.12 was asyncio
eventloop, but now we fail and not fall back to asyncio.
so all unittest on windows are failing on any import from
cassandra.connection.

in this change we use conan to download libev, and using
it to compile the driver with libev

Ref: https://conan.io/center/recipes/libev