fix: HostConnectionPool._maybe_trash_connection not to lock connection and pool
Locking connection a pool can lead to a deadlock.
Consider that connection.close also locks connection.
This commit also fixes condition for closing connection right away.
Instead of checking there is no in-flight requests it now checks if
in-flight requests less or equal than orphaned requests, since orphaned
requests still counts as in-flight and can be safely disregarded when connection
is dropped.