Replace self.assertRegex with new util function
This is part of migration effort to pytest. This commit was done by
hand. I introduced a new util function, and replaced the usages of
self.assertRegex with it. Why this way? If we ever need better error
messages for those asserts, we'll be able to do this by just modifying
this function, not all callsites.